pyproject.toml
[project]
name = "knowledgebase-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.24",
    "langchain-community>=0.3.22",
    "langchain-openai>=0.3.14",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.2",
    "streamlit>=1.44.1",
]
